Create a separate file for ColStats utils
